refactor: split enable/disable test helpers
