+++ addon-test-support/-private/disable-feature.js
@@ -1,0 +1,7 @@
+import { getContext } from '@ember/test-helpers';
+
+export function _disableFeature(featureName) {
+  let { owner } = getContext();
+  const featuresService = owner.lookup('service:features');
+  featuresService.disable(featureName);
+}
