+++ addon-test-support/-private/enable-feature.js
@@ -1,0 +1,4 @@
+export function _enableFeature(owner, featureName) {
+  const featuresService = owner.lookup('service:features');
+  featuresService.enable(featureName);
+}
