+++ addon-test-support/helpers/with-feature.js
@@ -1,5 +1,5 @@
 import { registerHelper } from '@ember/test';
-import { _enableFeature } from '../-private/toggle-feature';
+import { _enableFeature } from '../-private/enable-feature';
 
 registerHelper('withFeature', function withFeature(app, featureName) {
   _enableFeature(app.__deprecatedInstance__, featureName);
