+++ addon-test-support/index.js
@@ -1,12 +1,9 @@
 import { getContext } from '@ember/test-helpers';
-import { _enableFeature, _disableFeature } from './-private/toggle-feature';
+import { _enableFeature } from './-private/enable-feature';
 
 export function enableFeature(featureName) {
   let { owner } = getContext();
   _enableFeature(owner, featureName);
 }
 
-export function disableFeature(featureName) {
-  let { owner } = getContext();
-  _disableFeature(owner, featureName);
-}
+export { _disableFeature as disableFeature } from './-private/disable-feature';
